m4: Refactor evaluate module, and some modifications to improve nested evaluation
